requirements.txt
langchain==0.3.1
langchain-core==0.3.6
langchain-google-genai==0.1.8
langchain-community==0.3.1
langchain-huggingface==0.1.0
langgraph==0.2.28
langgraph-checkpoint==1.0.9
faiss-cpu==1.13.2
sentence-transformers==3.1.1
google-generativeai==0.7.2
streamlit==1.38.0
python-dotenv==1.0.1